README fixes
After cloning the repository, user should execute npm install instead of npm init in order to have the client or the server run properly. Update link to the server tutorial.Remove directory name parameter from the last git clone example to make it consistent with the rest of the README file.