improve handling of ExtensionObject in pure 1.04 format
In 1.04, the definition of extension Object is now attached the
corresponding dataType object.
In 1.03 the extension object definition was store into a embedded
BSD in a DataTypeSystemType. This required a complex preparation
on the client side.

DataType
  | 1
  |
  | n
  +-HasEncoding-> "Default Binary" (O)[DataTypeEncodingType]
                     |
                     +-HasDescription->"MyItemType" (V)[DataTypeDescriptionType]
                                         |
					 +-ComponentOf->Schema(V) []
                                              |
                                              +-ComponentOf->OPC Binary(V)[DataTypeSystemType]

Up to recently, 1.04 servers was mainly mixing the 1.03 and 1.04
approcah in the address space.
New 1.04 server are now only exposing dataType definition inside
the DataTypeDefinition attribute of the DataType Object and do not
expose the DataTypeDescriptionType nor the Schema nodes

In this commit, we introduce support for this new setup.