fix: code behaves differently when executed in `jest` runtime
When implementing tests using `jest` in projects that use this
library, tests fail for no apparent reason. This is because the
implementation of this library uses `promisify` of the node internal
`util` module to promisify some functions implemented using
callbacks. The callback functions created by `promisify` do not
fulfill the `instanceof Function` check when executed in a `jest`
environment which leads to unexpected errors during tests execution.
This issue is related to the following issue in the `jest` project
https://github.com/facebook/jest/issues/2549.
A simple workaround was to replace all `instanceof Function` checks
by a `typeof callback === 'function'` check.
The issue was described in the following comment:
https://github.com/facebook/jest/issues/2549#issuecomment-1104834577