Create dummy key using crypto iso parsing invalid key.
Since `invalidPrivateKey` is essentially only used to have a
value of the right type and to detect that the private key
has not been set by the API user, a random private key can
suffice as there is an insignificant chance there will ever
be a collision. Also creating a private key from value 0
is a bug magnet as was demonstrated by the Java TLS implementation
quite recently.

By creating a key, the key will be created by the same library
which will be using it, so the likelihood of subtle conflicts
between BoringSSL and OpenSSL becomes much more remote.